requirements.txt
# Databricks and Data Science Dependencies
databricks-connect>=0.9.0
pyspark>=3.4.0
delta-spark>=2.4.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
pyarrow>=10.0.0

# Machine Learning
scikit-learn>=1.2.0
mlflow>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
click>=8.1.0

# Development and Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
